@@ -5,7 +5,7 @@
     <OutputType>Exe</OutputType>
     <RootNamespace>Benchmarks</RootNamespace>
     <Configuration>Release</Configuration>
-    <TargetFrameworks>net472;netcoreapp2.1</TargetFrameworks>
+    <TargetFrameworks>net472;net5.0</TargetFrameworks>
     <RuntimeIdentifier>win7-x64</RuntimeIdentifier>
     <PlatformTarget>AnyCPU</PlatformTarget>
     <SignAssembly>false</SignAssembly>
@@ -13,18 +13,18 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\MiniProfiler.Shared\MiniProfiler.Shared.csproj" />
-    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
     <PackageReference Include="Jil" Version="2.16" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="System.Ben" Version="1.0.0" />
   </ItemGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp2.1'">
-    <PackageReference Include="ServiceStack.Text.Core" Version="5.4.0" />
+  <ItemGroup Condition="'$(TargetFramework)' == 'net5.0'">
+    <PackageReference Include="ServiceStack.Text.Core" Version="5.12.0" />
   </ItemGroup>
   
   <ItemGroup Condition="'$(TargetFramework)' == 'net472'">
-    <PackageReference Include="ServiceStack.Text" Version="5.4.0" />
+    <PackageReference Include="ServiceStack.Text" Version="5.12.0" />
     <Reference Include="System.Web.Extensions" />
   </ItemGroup>
 </Project>
