@@ -2,12 +2,13 @@
   <PropertyGroup>
     <AssemblyName>MiniProfiler.Tests.AspNetCore</AssemblyName>
     <RootNamespace>StackExchange.Profiling.Tests</RootNamespace>
-    <TargetFrameworks>net461;netcoreapp2.1;netcoreapp3.1</TargetFrameworks>
+    <TargetFrameworks>net461;netcoreapp3.1;net5.0</TargetFrameworks>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\MiniProfiler.Tests\MiniProfiler.Tests.csproj" />
     <ProjectReference Include="..\..\src\MiniProfiler.AspNetCore.Mvc\MiniProfiler.AspNetCore.Mvc.csproj" />
-    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="2.0.0" Condition="$(TargetFramework) != 'netcoreapp3.1'" />
+    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="2.0.0" Condition="$(TargetFramework) == 'net461'" />
     <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="3.1.3" Condition="$(TargetFramework) == 'netcoreapp3.1'" />
+    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="5.0.11" Condition="$(TargetFramework) == 'net5.0'" />
   </ItemGroup>
 </Project>
