@@ -1,9 +1,10 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Data.Sqlite;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Hosting;
 using Samples.AspNetCore.Models;
 using StackExchange.Profiling.Storage;
 
@@ -12,14 +13,15 @@ namespace Samples.AspNetCore
     public class Startup
     {
         public static string SqliteConnectionString { get; } = "Data Source=Samples; Mode=Memory; Cache=Shared";
+
         private static readonly SqliteConnection TrapConnection = new SqliteConnection(SqliteConnectionString);
 
-        public Startup(IHostingEnvironment env)
+        public Startup(IWebHostEnvironment env)
         {
             TrapConnection.Open(); //Hold the in-memory SQLite database open
 
             var builder = new ConfigurationBuilder()
-                .SetBasePath(env.ContentRootPath)
+                .SetBasePath(env.WebRootPath)
                 .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
                 .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true)
                 .AddEnvironmentVariables();
@@ -33,7 +35,12 @@ namespace Samples.AspNetCore
         {
             // Add framework services.
             services.AddDbContext<SampleContext>();
-            services.AddMvc();
+            services.AddMvc(options =>
+            {
+                // Because the samples have some MyAction and MyActionAsync duplicates
+                // See: https://github.com/aspnet/AspNetCore/issues/8998
+                options.SuppressAsyncSuffixInActionNames = false;
+            });
 
             // Add MiniProfiler services
             // If using Entity Framework Core, add profiling for it as well (see the end)
@@ -69,51 +76,73 @@ namespace Samples.AspNetCore
                 // The default handles async and works fine for almost all applications
                 //options.ProfilerProvider = new MyProfilerProvider();
 
+                // Optionally disable "Connection Open()", "Connection Close()" (and async variants).
+                //options.TrackConnectionOpenClose = false;
+
                 // Optionally use something other than the "light" color scheme.
                 options.ColorScheme = StackExchange.Profiling.ColorScheme.Auto;
 
-                // Optionally disable "Connection Open()", "Connection Close()" (and async variants).
-                //options.TrackConnectionOpenClose = false;
-                
+                // Enabled sending the Server-Timing header on responses
+                options.EnableServerTimingHeader = true;
+
+                // Optionally disable MVC filter profiling
+                //options.EnableMvcFilterProfiling = false;
+                // Or only save filters that take over a certain millisecond duration (including their children)
+                //options.MvcFilterMinimumSaveMs = 1.0m;
+
+                // Optionally disable MVC view profiling
+                //options.EnableMvcViewProfiling = false;
+                // Or only save views that take over a certain millisecond duration (including their children)
+                //options.MvcViewMinimumSaveMs = 1.0m;
+
+                // This enables debug mode with stacks and tooltips when using memory storage
+                // It has a lot of overhead vs. normal profiling and should only be used with that in mind
+                //options.EnableDebugMode = true;
+
                 // Optionally listen to any errors that occur within MiniProfiler itself
                 //options.OnInternalError = e => MyExceptionLogger(e);
-            }).AddEntityFramework();
 
-            services.AddLogging(builder =>
-            {
-                builder.AddConfiguration(Configuration.GetSection("Logging"));
-                builder.AddConsole();
-                builder.AddDebug();
-            });
+                options.IgnoredPaths.Add("/lib");
+                options.IgnoredPaths.Add("/css");
+                options.IgnoredPaths.Add("/js");
+            }).AddEntityFramework();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-                app.UseBrowserLink();
             }
             else
             {
                 app.UseExceptionHandler("/Home/Error");
             }
 
-            app.UseMiniProfiler();
 
-            app.UseStaticFiles();
+            app.UseMiniProfiler()
+               .UseStaticFiles()
+               .UseRouting()
+               .UseEndpoints(endpoints =>
+               {
+                   endpoints.MapAreaControllerRoute("areaRoute", "MySpace",
+                       "MySpace/{controller=Home}/{action=Index}/{id?}");
+                   endpoints.MapControllerRoute("default_route", "{controller=Home}/{action=Index}/{id?}");
 
-            app.UseMvc(routes =>
-            {
-                routes.MapRoute(
-                    name: "areaRoute",
-                    template: "{area:exists}/{controller=Home}/{action=Index}/{id?}");
+                   endpoints.MapRazorPages();
+                   endpoints.MapGet("/named-endpoint", async httpContext =>
+                   {
+                       var endpointName = httpContext.GetEndpoint().DisplayName;
+                       await httpContext.Response.WriteAsync($"Content from an endpoint named {endpointName}");
+                   }).WithDisplayName("Named Endpoint");
 
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
-            });
+                   endpoints.MapGet("implicitly-named-endpoint", async httpContext =>
+                   {
+                       var endpointName = httpContext.GetEndpoint().DisplayName;
+                       await httpContext.Response.WriteAsync($"Content from an endpoint named {endpointName}");
+                   });
+               });
 
             var serviceScopeFactory = app.ApplicationServices.GetRequiredService<IServiceScopeFactory>();
             using (var serviceScope = serviceScopeFactory.CreateScope())
