@@ -1,11 +1,13 @@
 ﻿using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
 using StackExchange.Profiling;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 
 namespace Benchmarks.Benchmarks
 {
-    [ClrJob, CoreJob]
+    [SimpleJob(RuntimeMoniker.Net472)]
+    [SimpleJob(RuntimeMoniker.Net50)]
     [Config(typeof(Configs.Memory))]
     public class DictionaryBenchmarks
     {
