@@ -1,8 +1,11 @@
 ﻿using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
 using StackExchange.Profiling;
+
 namespace Benchmarks
 {
-    [ClrJob, CoreJob]
+    [SimpleJob(RuntimeMoniker.Net472)]
+    [SimpleJob(RuntimeMoniker.Net50)]
     [Config(typeof(Configs.Memory))]
     public class MiniProfilerBenchmarks
     {
