@@ -1,10 +1,12 @@
 ﻿using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
 using StackExchange.Profiling;
 using StackExchange.Profiling.Internal;
 
 namespace Benchmarks
 {
-    [ClrJob, CoreJob]
+    [SimpleJob(RuntimeMoniker.Net472)]
+    [SimpleJob(RuntimeMoniker.Net50)]
     [Config(typeof(Configs.Memory))]
     public class SerializationBenchmarks
     {
